Move codebase from `test` to `internal` directory structure.
Refactored all references and file paths from `test` to `internal` to align with Go's best practices for internal package organization. Updated import statements, configuration paths, and related resources to reflect the new directory structure.
