+++ internal/doc/swagger-ui-example.png
