+++ internal/main.go
@@ -3,13 +3,13 @@ package main
 import (
 	log "github.com/sirupsen/logrus"
 	"github.com/softwareplace/go-password/pkg/str"
+	"github.com/softwareplace/goserve/internal/service/apiservice"
+	"github.com/softwareplace/goserve/internal/service/login"
+	"github.com/softwareplace/goserve/internal/service/provider"
 	"github.com/softwareplace/goserve/logger"
 	"github.com/softwareplace/goserve/security"
 	"github.com/softwareplace/goserve/security/secret"
 	"github.com/softwareplace/goserve/server"
-	"github.com/softwareplace/goserve/test/service/apiservice"
-	"github.com/softwareplace/goserve/test/service/login"
-	"github.com/softwareplace/goserve/test/service/provider"
 	"github.com/softwareplace/goserve/utils"
 	"os"
 )
@@ -25,7 +25,7 @@ func init() {
 
 		log.Infof("API_SECRET_KEY: %s", randomString)
 		_ = os.Setenv("API_SECRET_KEY", randomString)
-		_ = os.Setenv("API_PRIVATE_KEY", "./test/resource/secret/private.key")
+		_ = os.Setenv("API_PRIVATE_KEY", "./internal/resource/secret/private.key")
 	}
 
 }
@@ -52,7 +52,7 @@ func runSecretApi() {
 		SecurityService(securityService).
 		EmbeddedServer(apiservice.Register).
 		Get(apiservice.ReportCallerHandler, "/report/caller").
-		SwaggerDocHandler("./test/resource/pet-store.yaml").
+		SwaggerDocHandler("./internal/resource/pet-store.yaml").
 		StartServer()
 }
 
@@ -67,7 +67,7 @@ func runPublicApi() {
 	server.Default().
 		LoginService(loginService).
 		SecurityService(securityService).
-		SwaggerDocHandler("./test/resource/pet-store.yaml").
+		SwaggerDocHandler("./internal/resource/pet-store.yaml").
 		EmbeddedServer(apiservice.Register).
 		Get(apiservice.ReportCallerHandler, "/report/caller").
 		StartServer()
