+++ internal/resource/secret/private.key
