+++ internal/service/inventory/inventory_service.go
@@ -2,7 +2,7 @@ package inventory
 
 import (
 	goservectx "github.com/softwareplace/goserve/context"
-	"github.com/softwareplace/goserve/test/service/base"
+	"github.com/softwareplace/goserve/internal/service/base"
 	"net/http"
 	"sync"
 )
