+++ internal/service/login/login_service.go
@@ -3,13 +3,13 @@ package login
 import (
 	log "github.com/sirupsen/logrus"
 	goservectx "github.com/softwareplace/goserve/context"
+	"github.com/softwareplace/goserve/internal/service/provider"
 	"github.com/softwareplace/goserve/security"
 	"github.com/softwareplace/goserve/security/encryptor"
 	"github.com/softwareplace/goserve/security/jwt"
 	"github.com/softwareplace/goserve/security/login"
 	"github.com/softwareplace/goserve/security/model"
 	"github.com/softwareplace/goserve/security/principal"
-	"github.com/softwareplace/goserve/test/service/provider"
 	"github.com/softwareplace/goserve/utils"
 	"sync"
 	"time"
